docs: buttons components
