+++ src/components/EditButton.vue
@@ -1,0 +1,3 @@
+<template>
+  <button @click="$router.push('edit')">🛠️</button>
+</template>
