+++ src/views/EditorView.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>
+    <h1>Editor View Component</h1>
+  </div>
+</template>
+
+<style lang="css" scoped>
+div {
+  height: 50vh;
+  width: 96vw;
+}
+h1 {
+  color: black;
+}
+</style>
